refactor: move date input into separate component
